requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.2.3
bleach==6.2.0
boto3==1.34.136
botocore==1.34.136
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
cli_helpers==2.3.1
click==8.1.7
configobj==5.0.8
crashtest==0.4.1
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
docopt==0.6.2
dulwich==0.21.7
ecdsa==0.19.0
einops==0.8.0
esptool==4.7.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.3
frozenlist==1.4.1
fsspec==2024.6.0
future @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
huggingface-hub==0.25.1
idna==3.7
importlib_metadata==7.2.0
installer==0.7.0
intelhex==2.3.0
ipython==8.12.3
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==24.3.1
kornia==0.7.3
kornia_rs==0.1.5
macholib @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.24.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pgcli==4.1.0
pgspecial==2.1.2
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.47
propcache==0.2.0
psutil==6.0.0
psycopg==3.1.19
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyproject_hooks==1.1.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.3
reedsolo==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.5
scipy==1.13.1
sentencepiece==0.2.0
setproctitle==1.3.3
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/626bfa9c-8221-11ef-b9bf-daac7d5d70b1/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
soundfile==0.12.1
soupsieve==2.6
spandrel==0.4.0
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.13.0rc2
tabulate==0.9.0
tinycss2==1.4.0
tokenizers==0.20.0
tomli==2.0.1
tomlkit==0.12.5
torch==2.5.0.dev20240621
torchaudio==2.4.0.dev20240621
torchsde==0.2.6
torchvision==0.20.0.dev20240621
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trampoline==0.1.2
transformers==4.45.2
trove-classifiers==2024.5.22
typing_extensions==4.12.2
urllib3==1.26.19
virtualenv==20.26.2
wcwidth==0.2.13
webencodings==0.5.1
xattr==1.1.0
yarg==0.1.9
yarl==1.14.0
zipp==3.19.2
